сделал разбиение на модули как в вебинаре, чтобы сделать дз к уроку 6
