+++ app/build.gradle
@@ -30,6 +30,8 @@ dependencies {
     implementation project(Modules.core)
     implementation project(Modules.model)
     implementation project(Modules.utils)
+    implementation project(Modules.repository)
+    implementation project(Modules.historyScreen)
 
     //Kotlin
     implementation Kotlin.core
