--- app/src/main/java/geekbrains/ru/translator/model/datasource/ApiService.kt
@@ -1,13 +1,0 @@
-package geekbrains.ru.translator.model.datasource
-
-import geekbrains.ru.model.data.DataModel
-import kotlinx.coroutines.Deferred
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ApiService {
-
-    //метод теперь возвращает Deferred
-    @GET("words/search")
-    fun search(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
-}
