+++ buildSrc/src/main/java/Dependencies.kt
@@ -17,11 +17,11 @@ object Releases {
 object Modules {
    const val app = ":app"
    const val core = ":core"
-  const val model = ":model"
-//   const val repository = ":repository"
+   const val model = ":model"
+   const val repository = ":repository"
    const val utils = ":utils"
-//   //Features
-//   const val historyScreen = ":historyScreen"
+ //Features
+   const val historyScreen = ":historyScreen"
 }
 
 object Versions {
