+++ historyScreen/src/main/java/com/bartex/historyscreen/HistoryActivity.kt
@@ -1,0 +1,61 @@
+package com.bartex.historyscreen
+
+import android.os.Bundle
+import android.view.MenuItem
+import androidx.lifecycle.Observer
+import com.bartex.core2.BaseActivity
+import geekbrains.ru.model.data.AppState
+import geekbrains.ru.model.data.DataModel
+import kotlinx.android.synthetic.main.activity_history.*
+import org.koin.android.viewmodel.ext.android.viewModel
+
+class HistoryActivity : BaseActivity<AppState, HistoryInteractor>() {
+
+    override lateinit var model: HistoryViewModel
+    private val adapter: HistoryAdapter by lazy { HistoryAdapter() }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_history)
+
+        supportActionBar?.setHomeButtonEnabled(true)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+
+        iniViewModel()
+        initViews()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        model.getData("", false)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            android.R.id.home -> {
+                onBackPressed()
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
+    override fun setDataToAdapter(data: List<DataModel>) {
+        adapter.setData(data)
+    }
+
+    private fun iniViewModel() {
+        if (history_activity_recyclerview.adapter != null) {
+            throw IllegalStateException("The ViewModel should be initialised first")
+        }
+        val viewModel: HistoryViewModel by viewModel()
+        model = viewModel
+        model.getResult().observe(this@HistoryActivity, Observer<AppState> {
+            renderData(it)
+        })
+    }
+
+    private fun initViews() {
+        history_activity_recyclerview.adapter = adapter
+    }
+}
