+++ historyScreen/src/main/java/com/bartex/historyscreen/HistoryAdapter.kt
@@ -1,0 +1,47 @@
+package com.bartex.historyscreen
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.recyclerview.widget.RecyclerView
+import geekbrains.ru.model.data.DataModel
+import kotlinx.android.synthetic.main.history_item.view.*
+
+class HistoryAdapter : RecyclerView.Adapter<HistoryAdapter.RecyclerItemViewHolder>() {
+
+    private var data: List<DataModel> = arrayListOf()
+
+    fun setData(data: List<DataModel>) {
+        this.data = data
+        notifyDataSetChanged()
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerItemViewHolder {
+        return RecyclerItemViewHolder(
+            LayoutInflater.from(parent.context)
+                .inflate(R.layout.history_item, parent, false) as View
+        )
+    }
+
+    override fun onBindViewHolder(holder: RecyclerItemViewHolder, position: Int) {
+        holder.bind(data[position])
+    }
+
+    override fun getItemCount(): Int {
+        return data.size
+    }
+
+    inner class RecyclerItemViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+
+        fun bind(data: DataModel) {
+            if (layoutPosition != RecyclerView.NO_POSITION) {
+                itemView.tv_header_history.text = data.text
+                itemView.setOnClickListener {
+
+                    Toast.makeText(itemView.context, "on click: ${data.text}", Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
+    }
+}
