+++ historyScreen/src/main/res/values/colors.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="colorPrimary">#008577</color>
+    <color name="colorPrimaryDark">#00574B</color>
+    <color name="colorAccent">#D81B60</color>
+
+    <color name="color_error">#ef9a9a</color>
+    <color name="color_text_main">#66000014</color>
+    <color name="color_delete_text_icon">#66000014</color>
+    <color name="color_text_header">#313131</color>
+    <color name="color_black">#313131</color>
+    <color name="color_white">#FFFFFF</color>
+    <color name="color_white_transparent">#80FFFFFF</color>
+    <color name="color_gray">#66000000</color>
+    <color name="color_button_enabled">@color/colorAccent</color>
+    <color name="color_button_disabled">@color/color_gray</color>
+</resources>
