+++ repository/src/main/java/com/bartex/repository/api/ApiService.kt
@@ -1,0 +1,13 @@
+package com.bartex.repository.api
+
+import geekbrains.ru.model.data.DataModel
+import kotlinx.coroutines.Deferred
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ApiService {
+
+    //метод теперь возвращает Deferred
+    @GET("words/search")
+    fun search(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
+}
