+++ repository/src/main/java/com/bartex/repository/api/BaseInterceptor.kt
@@ -1,0 +1,50 @@
+package com.bartex.repository.api
+
+import okhttp3.Interceptor
+import okhttp3.Response
+import java.io.IOException
+import kotlin.jvm.Throws
+
+/**
+ * Custom interceptor to intercept basic responses and to show basic errors to the user
+ * (not fully implemented)
+ */
+class BaseInterceptor private constructor() : Interceptor {
+
+    private var responseCode: Int = 0
+
+    @Throws(IOException::class)
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val response = chain.proceed(chain.request())
+        responseCode = response.code()
+        return response
+    }
+
+    fun getResponseCode(): ServerResponseStatusCode {
+        var statusCode = ServerResponseStatusCode.UNDEFINED_ERROR
+        when (responseCode / 100) {
+            1 -> statusCode = ServerResponseStatusCode.INFO
+            2 -> statusCode = ServerResponseStatusCode.SUCCESS
+            3 -> statusCode = ServerResponseStatusCode.REDIRECTION
+            4 -> statusCode = ServerResponseStatusCode.CLIENT_ERROR
+            5 -> statusCode = ServerResponseStatusCode.SERVER_ERROR
+        }
+        return statusCode
+    }
+
+
+    enum class ServerResponseStatusCode {
+        INFO,
+        SUCCESS,
+        REDIRECTION,
+        CLIENT_ERROR,
+        SERVER_ERROR,
+        UNDEFINED_ERROR
+    }
+
+    companion object {
+
+        val interceptor: BaseInterceptor
+            get() = BaseInterceptor()
+    }
+}
